Moved controller to package controller.
